MCP stdio testing working
