+++ crates/hyperzoekt/Cargo.toml
@@ -33,7 +33,7 @@ notify = "6"
 toml = "0.7"
 sha2 = "0.10"
 signal-hook = "0.3"
-rust-mcp-sdk = { version = "0.6", features = ["server"] }
+rust-mcp-sdk = { version = "0.6", features = ["server", "client"] }
 rust-mcp-transport = { version = "0.5" }
 hyper = { version = "0.14", features = ["full"] }
 tokio-stream = "0.1"
@@ -47,6 +47,6 @@ predicates = "2"
 
 [[bin]]
 name = "hyperzoekt"
-path = "src/bin/hyperzoekt.rs"
+path = "src/bin/hyperzoekt/main.rs"
